self_chatbot/backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# HTTP and API clients
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Azure OpenAI
openai==1.6.1
azure-identity==1.15.0

# Google Cloud (Gemini)
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2

# AWS (Claude)
boto3==1.34.0
botocore==1.34.0

# Image processing
Pillow==10.1.0

# Environment and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
curl-cffi==0.6.2

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0

# JSON and data handling
orjson==3.9.10

# Additional utilities
validators==0.22.0